+++ graph/graph-application/src/main/kotlin/eu/tib/orkg/prototype/statements/api/compat.kt
@@ -36,6 +36,8 @@ interface ListUseCases : CreateListUseCase, RetrieveListUseCase, UpdateListUseCa
 
 interface ClassHierarchyUseCases : CreateClassHierarchyUseCase, RetrieveClassHierarchyUseCase, DeleteClassHierarchyUseCase
 
+interface ResearchFieldHierarchyUseCases : RetrieveResearchFieldHierarchyUseCase
+
 // FIXME: we need to refactor those as well
 interface OtherResourceUseCases :
     MarkAsVerifiedUseCase,