+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added new endpoints to fetch research field hierarchies
+  (See: [!606](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/606))
 ### Fixed
 - Lists that still have elements can now be deleted.
   (See: [!627](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/627))